moved libs & store directories
