+++ src/libs/hash.ts
